requirements.txt
# Dependencies pour Growth Weekly Newsletter Generator

# Core dependencies
python-dotenv>=1.0.0

# Gmail API
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# AI / Anthropic
anthropic>=0.18.0

# Configuration
PyYAML>=6.0.0

# Optional: For better performance
requests>=2.31.0

# Core Python (requires Python 3.11+)
python-dotenv==1.0.0

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0

# AI/LLM APIs
anthropic==0.18.1
openai==1.12.0  # Alternative si préféré

# Data processing
pandas==2.2.0
python-dateutil==2.8.2

# Configuration and utilities
pyyaml==6.0.1
tqdm==4.66.2  # Progress bars
loguru==0.7.2  # Better logging

# Date and time
pytz==2024.1

# JSON handling
orjson==3.9.15  # Faster JSON parsing

# Type hints and validation (optional but recommended)
pydantic==2.6.1
